Update Lib_LCD.c
Added in actual code into the following functions:
vLCD_GO_TO_POSITION()
VLCD_HOME_TOP_LINE()
vLCD_HOME_BOTTOM_LINE()

Assumption was made that when doing a home the cursor is just going to be moved to the beginning of each line. For the top line, this was easily done using one of the built in commands for the LCD. Homing the bottom line is done by calling the position function to move the cursor.